refactor: simplify file structure
