+++ src/admin/actions/acceptGroupApply.js
@@ -1,7 +1,6 @@
 import { Group, GroupApply, User } from "../../db";
 import { logger } from "../../utils/logger";
 import { sendApplyDoneMessage } from "../../utils/slack";
-import acceptGroup from "../api/acceptGroup";
 
 export const acceptGroupAction = {
   actionType: "record",
@@ -18,4 +17,37 @@ export const acceptGroupAction = {
       msg: "Accepting Group: Success",
     };
   },
+};
+
+const acceptGroup = async (currAdmin, groupName) => {
+  const adminName = currAdmin.user.username;
+  logger.admin.info(
+    "Accept Group; name: %s, adminUser: %s",
+    groupName,
+    adminName,
+  );
+
+  const newGroup = await GroupApply.findOneAndDelete({ name: groupName });
+  const newGroupJSON = newGroup.toJSON({ virtuals: false });
+  delete newGroupJSON._id;
+  newGroupJSON.members.forEach((member, i) => {
+    delete newGroupJSON.members[i]._id;
+  });
+  const created = await Group.create(newGroupJSON);
+
+  currAdmin.actionHistory.push({
+    name: "acceptGroup",
+    target: created._id,
+  });
+
+  await Promise.all([
+    ...newGroupJSON.members.map(({ user }) =>
+      User.findByIdAndUpdate(user._id, { $push: { groups: created._id } }),
+    ),
+    currAdmin.save(),
+  ]);
+
+  sendApplyDoneMessage(groupName, currAdmin);
+
+  return created;
 };
